+++ src/extensions/rules/rulesext.cpp
@@ -102,7 +102,10 @@ RulesClassExtension::RulesClassExtension(const RulesClass* this_ptr) :
     IsAIDetectDisguise(true),
     IsAIOneHarvesterInSingleplayer(true),
     PausedRepairsFrame(6),
-    BaseUnit()
+    BaseUnit(),
+    Diff(),
+    PlayerNormal(),
+    IsHasPlayerNormal(false)
 {
     //if (this_ptr) EXT_DEBUG_TRACE("RulesClassExtension::RulesClassExtension - 0x%08X\n", (uintptr_t)(ThisPtr));
 
@@ -381,7 +384,6 @@ void RulesClassExtension::Process(CCINIClass &ini)
      */
     Objects(ini);
 
-    This()->Difficulty(ini);
     This()->CrateRules(ini);
     This()->CombatDamage(ini);
     This()->AudioVisual(ini);
@@ -408,6 +410,7 @@ void RulesClassExtension::Process(CCINIClass &ini)
      *  #NOTE: These must be performed last!
      */
     General(ini);
+    Difficulty(ini);
     MPlayer(ini);
     AudioVisual(ini);
     CombatDamage(ini);
@@ -740,6 +743,8 @@ bool RulesClassExtension::General(CCINIClass &ini)
      */
     BaseUnit = TGet_TypeList(ini, GENERAL, "BaseUnit", BaseUnit);
 
+    IsHasPlayerNormal = ini.Get_Bool(GENERAL, "HasPlayerNormal", IsHasPlayerNormal);
+
     return true;
 }
 
@@ -1054,6 +1059,26 @@ bool RulesClassExtension::PrerequisiteGroups(CCINIClass& ini)
     }
 
     return counter > 0;
+}
+
+
+/**
+ *  Fetch the various difficulty group settings.
+ *
+ *  @author: Rampastring
+ */
+bool RulesClassExtension::Difficulty(CCINIClass& ini)
+{
+    Difficulty_Get(ini, Diff[DIFF_EASY], "Easy");
+    Difficulty_Get(ini, Diff[DIFF_NORMAL], "Normal");
+    Difficulty_Get(ini, PlayerNormal, "PlayerNormal");
+    Difficulty_Get(ini, Diff[DIFF_HARD], "Difficult");
+    Difficulty_Get(ini, Diff[EXT_DIFF_VERY_HARD], "VeryEasy");
+    Difficulty_Get(ini, Diff[EXT_DIFF_BRUTAL], "BrutallyEasy");
+    Difficulty_Get(ini, Diff[EXT_DIFF_EXTREME], "ExtremelyEasy");
+    Difficulty_Get(ini, Diff[EXT_DIFF_ULTIMATE], "UltimatelyEasy");
+
+    return true;
 }