+++ src/spawner/spawner.h
@@ -40,14 +40,16 @@ class Spawner
 public:
     Spawner() = delete;
 
-    static void Init();
+    static bool Init();
     static bool Start_Game();
 
 private:
     static bool Start_Scenario(char* scenario_name);
     static bool Load_Game(const char* file_name);
 
-    static void Init_Session(char* scenario_name);
+    static int Spawner_Config_AI_Difficulty_To_Game_AI_Difficulty(int difficulty);
+
+    static bool Init_Session(char* scenario_name);
     static void Init_Network();
     static bool Reconcile_Players();